+++ CHANGELOG.md
@@ -1,16 +1,22 @@
 
 ### 4.0.11 (2020-07-23)
-##### 【bugfix】
+##### g2依赖 升级
+- g2@2.0.14
+- g2Plot@1.1.15
+##### bugfix
 - react-reconciler required but not add into deps #1219
 - Duplicate export #1221
 - 修复Coordinate组件的action属性
+##### feature
+- 图表联动，当前支持没有设置adjust的图表联动。详细见Tooltip linkage 属性文档。
+- Tooltip 点击后锁定位置。和Tooltip 点击后才展示。
 
 
 ### 4.0.10 (2020-07-14)
 ##### g2依赖 升级
 - g2@2.0.12
 - g2Plot@1.1.7 
-##### 【bugfix】
+##### bugfix
 - Roboto字体带来的,中文裁切问题 
 - 适配 g2 新增的一层div，dom 结构变化影响自定义Tooltip定位。